Write from PHP engine to Go context directly, using an exported callback.
Previously, the PHP engine would write to a temporary buffer which was
periodically synched on the Go side. This caused various issues (access
contention, delayed sync etc) which were resolved by having PHP write
to the Go context directly, via an exported function.

This requires us passing a pointer to the parent Go `php.Context`
structure, including the `io.Writer` embedded within, to the C context,
which is then passed back to Go for use as a function parameter in a
declared callback.